Assignment2/requirements.txt
arrow==1.2.3
black==22.12.0
click==8.1.3
dnspython==2.2.1
email-validator==1.1.2
Flask==2.2.2
fpdf==1.7.2
greenlet==2.0.1
idna==3.4
importlib-metadata==5.2.0
isbnlib==3.10.3
iso3166==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
make==0.1.6.post2
MarkupSafe==2.1.1
marshmallow==3.19.0
mypy-extensions==0.4.3
packaging==22.0
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.0
py-moneyed==0.8.0
pycountry==22.3.5
pypng==0.20220715.0
PyQRCode==1.2.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2020.1
reportlab==3.6.12
schwifty==2020.9.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
zipp==3.11.0
